+++ apps/mobile/src/ui/components/index.ts
@@ -7,6 +7,7 @@ export * from "./ConfirmationDialog";
 export * from "./EmptyState";
 export * from "./ErrorState";
 export * from "./FormField";
+export * from "./LedgerLoader";
 export * from "./MoneyValue";
 export * from "./OfflineBanner";
 export * from "./SelectionField";